Use CustomActionOperationMixin to centralize shared methods
